ap - Setup graph building based on database
